# gen/http/openapi3.yaml
openapi: 3.0.3
info:
    title: Calculator Service
    description: A simple calculator service
    version: 0.0.1
servers:
    - url: http://localhost:8080
paths:
    /add/{a}/{b}:
        get:
            tags:
                - calc
            summary: add calc
            description: Adds two integers and returns the result.
            operationId: calc#add
            parameters:
                - name: a
                  in: path
                  required: true
                  schema:
                    type: integer
                    example: 292483880
                    format: int64
                  example: 312340699
                - name: b
                  in: path
                  required: true
                  schema:
                    type: integer
                    example: 396407178
                    format: int64
                  example: 1242178904
            responses:
                "200":
                    description: OK response.
                    content:
                        application/json:
                            schema:
                                type: integer
                                example: 621720987
                                format: int64
                            example: 1496178131
    /sub/{a}/{b}:
        get:
            tags:
                - calc
            summary: subtract calc
            description: Subtracts two integers and returns the result.
            operationId: calc#subtract
            parameters:
                - name: a
                  in: path
                  required: true
                  schema:
                    type: integer
                    example: 1516744731
                    format: int64
                  example: 842555000
                - name: b
                  in: path
                  required: true
                  schema:
                    type: integer
                    example: 955595836
                    format: int64
                  example: 499160208
            responses:
                "200":
                    description: OK response.
                    content:
                        application/json:
                            schema:
                                type: integer
                                example: 1975077685
                                format: int64
                            example: 818051480
components: {}
tags:
    - name: calc
      description: The calc service provides addition.